requirements.txt
python-dotenv
O365